New project and some revision to how the lists of data are handled for models/meshes - New MORPG project added - Meshes now fully utilitize the power of the List! - The bool conversion operator of List is now removed (thank god!) as it was causing severe brain damage and loads of run-time errors. - GameState now re-named to AppState, as that fits it much better. - Includes adjusted for Mesh so that it isn't included by accident into every f****** file. Remember: Always use forward declarations where possible in header files, and include the actual relevant file in the implementation...

+++ src/System/Memory.cpp
@@ -8,7 +8,7 @@
 
 struct MemoryStatus 
 {
-	int freeBytes;
+	int64 freeBytes;
 	// Percentage.
 	int memoryLoad;
 } memoryStatus;
